Separated views for admin and user for a hackathon
